SSTrans_SBIR/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy
pillow
opencv-python
pyyaml
scikit-learn
matplotlib
tqdm
einops
scipy
faiss-cpu
